auto_test_platform/requirements.txt
# FastAPI 核心
fastapi==0.109.2
uvicorn[standard]==0.27.1

# 数据库
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.19.0

# Pydantic
pydantic==2.6.1

# 请求处理
httpx==0.26.0

# pytest 执行
pytest==8.0.0
pytest-asyncio==0.23.4
requests==2.31.0

# Allure 报告
allure-pytest==2.13.2

# 定时任务
APScheduler==3.10.4

# 其他工具
python-multipart==0.0.9
PyYAML==6.0.1
